e2e_tests/requirements.txt
# E2E Testing Dependencies for Playwright
#
# Install with: pip install -r e2e_tests/requirements.txt
# Then run: playwright install

# Core Playwright dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-playwright>=0.4.3

# Optional: For parallel test execution
pytest-xdist>=3.3.1

# Optional: For better test reporting
pytest-html>=3.2.0
